[refactor] Add test snapshots
